Feat Add: CORS MultiParser and GeneralParser
