+++ backend/src/main/java/org/scoula/safety_inspection/infra/cors/dto/CopyOfRegisterDto.java
@@ -8,21 +8,19 @@ import lombok.Setter;
 public class CopyOfRegisterDto {
 
     private Integer analysisNo;
-    private final String ownerState;
+    private final Double ownerState;
     private final String ownership;
     private final String commonOwner;
     private final String changeOwnerCount;
-    private final String maximumOfBond;
-    private final String deposit;
+    private final Integer maximumOfBond;
 
-    public CopyOfRegisterDto(Integer analysisNo,String ownerState, String ownership, String commonOwner, String changeOwnerCount, String maximumOfBond, String deposit) {
+    public CopyOfRegisterDto(Integer analysisNo,Double ownerState, String ownership, String commonOwner, String changeOwnerCount, Integer maximumOfBond) {
         this.analysisNo = analysisNo;
         this.ownerState = ownerState;
         this.ownership = ownership;
         this.commonOwner = commonOwner;
         this.changeOwnerCount = changeOwnerCount;
         this.maximumOfBond = maximumOfBond;
-        this.deposit = deposit;
     }
 
 }
