--- backend/src/main/resources/application.properties
@@ -1,13 +1,0 @@
-#driver=com.mysql.cj.jdbc.Driver
-#url=jdbc:mysql://127.0.0.1:3306/scoula_db
-jdbc.driver=net.sf.log4jdbc.sql.jdbcapi.DriverSpy
-jdbc.url=jdbc:log4jdbc:mysql://localhost:3307/estate
-jdbc.username=root
-jdbc.password=1234
-
-codef.eprepayNo=J77887546198
-codef.eprepayPass=aa1234
-codef.userPass=1234
-codef.userName=\uACE0\uC7AC\uC131
-codef.birthDate=971104
-codef.identity=1300111
